Moving deepcopy to  utils Exporting tenant type from azurecore
